finish refactor of request, but not vehicle
