+++ pyrai/__init__.py
@@ -1,2 +1,2 @@
-from .dispatcher.api import Fleet, Pyrai, FleetParams, Vehicle, VehicleEvent, UserKey, Location, Request, Event, Notification, NotificationData, StatusResponse, VehicleAssignments, StatusError, Metrics
+from .dispatcher.structures import Fleet, Pyrai, FleetParams, Vehicle, VehicleEvent, UserKey, Location, Request, Event, Notification, NotificationData, StatusResponse, VehicleAssignments, StatusError, Metrics
 __all__ = ["Fleet", "Pyrai", "FleetParams", "Vehicle", "VehicleEvent", "UserKey", "Location", "Request", "Event", "Notification", "NotificationData", "StatusResponse", "VehicleAssignments", "StatusError", "Metrics"]
