--- pyrai/dispatcher/api.py
@@ -1,6 +1,0 @@
-import requests
-import datetime
-import json
-from dateutil.parser import isoparse
-from .structures import *
-from pyrai.helpers import to_rfc3339
