+++ pyrai/dispatcher/methods/request/__init__.py
@@ -1,0 +1,2 @@
+from .cancel import cancel
+__all__ = ["cancel"]
